fix(amdsmi): share one bounded copy and guard zero-length buffers
- Move CopyBounded out of the container-ID parser header into
  amd_smi_utils.h beside the other string helpers, and have
  smi_clear_char_and_reinitialize delegate to it. Both carried the same
  min(capacity - 1, length) clamp and explicit NUL write.
- Guard a zero length in smi_clear_char_and_reinitialize. Its len
  parameter is uint32_t, so len - 1 wrapped to 4294967295 before the
  clamp and the copy ran with the full source length into a zero-byte
  buffer. No current call site passes zero, so this was latent.
- Cover a capacity of one, which only leaves room for the terminator.
